+++ src/main/java/cn/yuqi/mini/spring/beans/factory/BeansException.java
@@ -1,0 +1,11 @@
+package cn.yuqi.mini.spring.beans.factory;
+
+public class BeansException extends RuntimeException {
+    public BeansException(String message) {
+        super(message);
+    }
+
+    public BeansException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}