+++ src/main/java/cn/yuqi/mini/spring/beans/factory/config/SingletonBeanRegistry.java
@@ -1,4 +1,4 @@
-package cn.yuqi.mini.spring.beans.factory;
+package cn.yuqi.mini.spring.beans.factory.config;
 
 //负责 管理已经创建完成的单例 Bean 实例
 public interface SingletonBeanRegistry {