+++ src/test/java/cn/yuqi/mini/spring/beans/factory/BeanFactoryTest.java
@@ -1,8 +1,8 @@
 package cn.yuqi.mini.spring.beans.factory;
 
+import cn.yuqi.mini.spring.beans.factory.config.BeanDefinition;
+import cn.yuqi.mini.spring.beans.factory.support.DefaultListableBeanFactory;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class BeanFactoryTest {
 
@@ -17,14 +17,11 @@ class BeanFactoryTest {
         helloService.sayHello();
     }
 
-    @Test
-    void testGetBeans() {
-
-    }
-    static class HelloService {
-        public String sayHello() {
-            System.out.println("hello");
-            return "hello";
-        }
-    }
+    //此时如果使用下面这个类内定义的类进行测试会报错,需要一个顶级类
+//    static class HelloService {
+//        public String sayHello() {
+//            System.out.println("hello");
+//            return "hello";
+//        }
+//    }
 }